+++ summoner-cli/examples/cabalScript.hs
@@ -1,0 +1,8 @@
+#!/usr/bin/env cabal
+{- cabal:
+build-depends:
+  , base ^>= 4.15.1.0
+-}
+
+main :: IO ()
+main = putStrLn "Hello, World!"