+++ summoner-cli/golden/Main.hs
@@ -1,0 +1,9 @@
+module Main (main) where
+
+import Summoner.Golden (generateProjects, generateScripts, generateTomlConfig)
+
+main :: IO ()
+main = do
+    generateScripts
+    generateTomlConfig
+    generateProjects