+++ summoner-cli/src/Summoner/GhcVer.hs
@@ -1,6 +1,6 @@
 {- |
 Module                  : Summoner.GhcVer
-Copyright               : (c) 2017-2021 Kowainik
+Copyright               : (c) 2017-2022 Kowainik
 SPDX-License-Identifier : MPL-2.0
 Maintainer              : Kowainik <xrom.xkov@gmail.com>
 Stability               : Stable
@@ -24,8 +24,6 @@ module Summoner.GhcVer
     , oldGhcs
     ) where
 
-import Relude.Extra.Enum (inverseMap, universe)
-
 import qualified Data.Text as T
 import qualified Text.Show as Show
 
@@ -37,7 +35,8 @@ data GhcVer
     | Ghc844
     | Ghc865
     | Ghc884
-    | Ghc8103
+    | Ghc8107
+    | Ghc902
     deriving stock (Eq, Ord, Show, Enum, Bounded)
 
 -- | Converts 'GhcVer' into dot-separated string.
@@ -48,7 +47,8 @@ showGhcVer = \case
     Ghc844  -> "8.4.4"
     Ghc865  -> "8.6.5"
     Ghc884  -> "8.8.4"
-    Ghc8103 -> "8.10.3"
+    Ghc8107 -> "8.10.7"
+    Ghc902  -> "9.0.2"
 
 {- | These are old GHC versions that are not working with default GHC versions
 when using Stack.
@@ -66,8 +66,9 @@ latestLts = \case
     Ghc822  -> "lts-11.22"
     Ghc844  -> "lts-12.26"
     Ghc865  -> "lts-14.27"
-    Ghc884  -> "lts-16.17"
-    Ghc8103 -> "lts-17.0"
+    Ghc884  -> "lts-16.31"
+    Ghc8107 -> "lts-18.21"
+    Ghc902  -> "nightly-2022-01-10"
 
 -- | Represents PVP versioning (4 numbers).
 data Pvp = Pvp
@@ -89,7 +90,8 @@ baseVerPvp = \case
     Ghc844  -> Pvp 4 11 1 0
     Ghc865  -> Pvp 4 12 0 0
     Ghc884  -> Pvp 4 13 0 0
-    Ghc8103 -> Pvp 4 14 1 0
+    Ghc8107 -> Pvp 4 14 3 0
+    Ghc902  -> Pvp 4 15 1 0
 
 -- | Returns corresponding @base@ version of the given GHC version.
 baseVer :: GhcVer -> Text
@@ -106,8 +108,8 @@ baseVer = show . baseVerPvp
 -}
 cabalBaseVersions :: [GhcVer] -> Text
 cabalBaseVersions ghcVers = case sort ghcVers of
-    [] -> ""
-    [v] -> "^>= " <> baseVer v
+    []          -> ""
+    [v]         -> "^>= " <> baseVer v
     minGhc:x:xs -> ">= " <> baseVer minGhc <> " && < " <> upperBound (x :| xs)
   where
     upperBound :: NonEmpty GhcVer -> Text