+++ summoner-cli/src/Summoner/Source.hs
@@ -1,6 +1,6 @@
 {- |
 Module                  : Summoner.Source
-Copyright               : (c) 2017-2021 Kowainik
+Copyright               : (c) 2017-2022 Kowainik
 SPDX-License-Identifier : MPL-2.0
 Maintainer              : Kowainik <xrom.xkov@gmail.com>
 Stability               : Stable