+++ summoner-cli/test/Test/TomlSpec.hs
@@ -3,7 +3,6 @@ module Test.TomlSpec
        ) where
 
 import Hedgehog (MonadGen, forAll, tripping)
-import Relude.Extra.Enum (universe)
 import Test.Hspec (Spec, describe, it, shouldReturn, shouldSatisfy)
 import Test.Hspec.Hedgehog (hedgehog)
 import Toml.Codec.Code (decode, encode)